app/src/main/java/com/revdex/crypto/domain/use_case/get_coins/GetCoinsUseCase.kt
package com.revdex.crypto.domain.use_case.get_coins

import com.revdex.crypto.common.Resource
import com.revdex.crypto.domain.model.Coin
import com.revdex.crypto.domain.repository.CoinRepository
import kotlinx.coroutines.flow.flow
import retrofit2.HttpException
import java.io.IOException
import javax.inject.Inject

class GetCoinsUseCase @Inject constructor(
    private val repository: CoinRepository
) {
    operator fun invoke() = flow<Resource<List<Coin>>> {
        try {
            emit(Resource.Loading())
            val coins = repository.getCoins()
            emit(Resource.Success(coins))
        } catch (e: HttpException) {
            emit(Resource.Error(e.localizedMessage ?: "An unexpected error occured"))
        } catch (e: IOException) {
            emit(Resource.Error("Couldn't reach server. Check your internet connection."))
        } catch (e: Exception) {
            emit(Resource.Error("Generic error occurred"))
        }
    }
}